requirements.txt
pandas==2.2.2
pyarrow==17.0.0
python-dotenv==1.0.1
datasets==3.0.1
polars==1.8.2
pyspark==3.5.1

boto3==1.35.36
botocore==1.35.36
